feat: #103 - useGetSpace 훅으로 비즈니스 로직 분리
